+++ stories/src/simple/simple-scrollable.tsx
@@ -1,66 +1,88 @@
 import React, { Component } from 'react';
-import { DragDropContext, Droppable, Draggable } from '../../../src';
+import {
+  DragDropContext,
+  Droppable,
+  Draggable,
+  DraggableStyle,
+  DropResult,
+} from '../../../src';
 
 // fake data generator
-const getItems = (count) =>
-  // eslint-disable-next-line no-restricted-syntax
+const getItems = (count: number) =>
   Array.from({ length: count }, (v, k) => k).map((k) => ({
     id: `item-${k}`,
     content: `item ${k}`,
   }));
 
 // a little function to help us with reordering the result
-const reorder = (list, startIndex, endIndex) => {
-  const result = [...list];
+const reorder = <TList extends unknown[]>(
+  list: TList,
+  startIndex: number,
+  endIndex: number,
+): TList => {
+  const result = Array.from(list) as TList;
   const [removed] = result.splice(startIndex, 1);
   result.splice(endIndex, 0, removed);
 
   return result;
 };
 
-const withAssortedSpacing = () => ({
-  // margin
-  marginTop: 10,
-  // not allowing margin collapsing
-  // marginBottom: 20,
-  marginLeft: 30,
-  marginRight: 40,
-  // padding
-  paddingTop: 10,
-  paddingBottom: 20,
-  paddingLeft: 30,
-  paddingRight: 40,
-  // border
-  borderStyle: 'solid',
-  borderColor: 'purple',
-  borderTopWidth: 2,
-  borderBottomWidth: 4,
-  borderLeftWidth: 6,
-  borderRightWidth: 8,
-});
+const grid = 8;
 
-const getItemStyle = (isDragging, draggableStyle) => ({
+const getItemStyle = (
+  isDragging: boolean,
+  draggableStyle: DraggableStyle = {},
+) => ({
   // some basic styles to make the items look a bit nicer
-  userSelect: 'none',
-  ...withAssortedSpacing(),
+  userSelect: 'none' as const,
+  padding: grid * 2,
+  margin: `0 0 ${grid}px 0`,
+  border: '5px solid yellow',
+  height: 30,
 
   // change background colour if dragging
-  background: isDragging ? 'lightgreen' : 'pink',
+  background: isDragging ? 'lightgreen' : 'red',
 
   // styles we need to apply on draggables
   ...draggableStyle,
 });
 
-export default class App extends Component {
-  constructor(props, context) {
-    super(props, context);
+const getListStyle = (isDraggingOver: boolean, overflow?: string) => ({
+  background: isDraggingOver ? 'lightblue' : 'grey',
+  padding: grid,
+  border: '5px solid pink',
+  width: 250,
+  maxHeight: '50vh',
+  overflow,
+});
+
+interface AppProps {
+  overflow?: string;
+}
+
+interface Item {
+  id: string;
+  content: string;
+}
+
+interface AppState {
+  items: Item[];
+}
+
+export default class App extends Component<AppProps, AppState> {
+  static defaultProps = {
+    overflow: 'auto',
+  };
+
+  constructor(props: AppProps) {
+    super(props);
     this.state = {
       items: getItems(10),
     };
     this.onDragEnd = this.onDragEnd.bind(this);
   }
 
-  onDragEnd(result) {
+  onDragEnd(result: DropResult) {
     // dropped outside the list
     if (!result.destination) {
       return;
@@ -83,17 +105,17 @@ export default class App extends Component {
     return (
       <DragDropContext onDragEnd={this.onDragEnd}>
         <Droppable droppableId="droppable">
-          {(droppableProvided) => (
+          {(droppableProvided, droppableSnapshot) => (
             <div
               ref={droppableProvided.innerRef}
-              style={{
-                width: 250,
-                background: 'lightblue',
-
-                ...withAssortedSpacing(),
-                // no margin collapsing
-                marginTop: 0,
-              }}
+              style={getListStyle(
+                droppableSnapshot.isDraggingOver,
+                this.props.overflow,
+              )}
+              onScroll={(e) =>
+                // eslint-disable-next-line no-console
+                console.log('current scrollTop', e.currentTarget.scrollTop)
+              }
             >
               {this.state.items.map((item, index) => (
                 <Draggable key={item.id} draggableId={item.id} index={index}>
